+++ CHANGELOG.md
@@ -7,14 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 _No unreleased changes_
 
-## [3.2.0] - 2025-07-06
+## [3.2.0] - 2025-08-04
 ### Added 
 - Added twitch rewards queue link
 - Added display names to soundboard buttons
+- Added search bar to sounds 
 
 ### Changed
 - Changed channel point redemptions to use EventSub
 - Changed soundboard button sizes
+- Changed all sound rewards to use arrays 
+- Single click selects multiple sounds by default
 
 ## [3.1.0] - 2023-02-03
 ### Fixed